requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
booru==1.0.21
Brotli==1.1.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
curl_cffi==0.6.3
discord.py==2.3.2
discordLevelingSystem==1.2.1
distro==1.9.0
dnspython==2.6.1
duckduckgo_search==5.3.0
exceptiongroup==1.2.1
frozenlist==1.4.1
fuzzywuzzy==0.18.0
g4f==0.3.0.8
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
langdetect==1.0.9
Levenshtein==0.25.1
motor==3.4.0
multidict==6.0.5
nest-asyncio==1.6.0
openai==1.25.1
orjson==3.10.2
praw==7.7.1
prawcore==2.4.0
psutil==5.9.8
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.1
PyNaCl==1.5.0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pyttsx3==2.90
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.0
redis==5.0.4
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
spotipy==2.23.0
tqdm==4.66.4
typing_extensions==4.11.0
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.8.0
xmltodict==0.13.0
yarl==1.9.4